Fixes #77 Email notification are sent for projects updates when
using the built-in change_status method. When manually changing
the status of of project notifications will not be sent. It was
designed this way so that notification are sent when
a change in status occurs by comparing new_status to instance
status. This could also be accomplished by overinding the save
method but a database query would need to be perform (to check
for the old status and compare to new one) each time an instance
is saved.